+++ itachallenge-challenge/src/test/java/com/itachallenge/submission/document/SubmissionDocumentTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.time.LocalDateTime;
 import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -31,7 +32,7 @@ class SubmissionDocumentTest {
 
     @BeforeEach
     void setUp() {
-        submissionDocumentAllArgs = new SubmissionDocument(submissionId, userId, challengeId, languageId, submissionStatus, submissionText);
+        submissionDocumentAllArgs = new SubmissionDocument(submissionId, userId, challengeId, languageId, submissionStatus, submissionText, LocalDateTime.now());
         submissionDocumentNoArgs = new SubmissionDocument();
     }